Cleaned up and added table name customization
