+++ hedera-mirror-rest/entity-info-update/entityUpdateHandler.js
@@ -21,8 +21,6 @@
 'use strict';
 
 // external libraries
-const proto = require('@hashgraph/proto');
-const {PublicKey, KeyList} = require('@hashgraph/sdk');
 const log4js = require('log4js');
 
 // local
@@ -32,69 +30,20 @@ const utils = require('./utils');
 
 const logger = log4js.getLogger();
 
-const keyListToProto = (keyList) => {
-  const keys = [];
-
-  for (const key of keyList) {
-    keys.push(keyToProto(key));
-  }
-
-  return {
-    keys,
-  };
-};
-
-const keyToProto = (key) => {
-  if (key instanceof PublicKey) {
-    return {
-      ed25519: key.toBytes(),
-    };
-  }
-
-  if (key instanceof KeyList) {
-    return {
-      keyList: keyListToProto(key),
-    };
-  }
-
-  throw Error('Unsupported key type');
-};
-
-const getProtoAndEd25519HexFromPublicKey = (key) => {
-  const protoKey = {
-    ed25519: key.toBytes(),
-  };
-  const ed25519Hex = key.toString();
-
-  return {protoKey, ed25519Hex};
-};
-
-const getProtoAndEd25519HexFromKeyList = (key) => {
-  const protoKey = {
-    keyList: keyListToProto(key),
-  };
-  const ed25519Hex = key._keys.toString().split(',')[0];
-
-  return {protoKey, ed25519Hex};
-};
-
-const getBufferAndEd25519HexFromKey = (key) => {
-  const {protoKey, ed25519Hex} =
-    key instanceof PublicKey ? getProtoAndEd25519HexFromPublicKey(key) : getProtoAndEd25519HexFromKeyList(key);
-
-  let protoBuffer = null;
-  protoBuffer = proto.Key.encode(protoKey, protoBuffer).finish();
-  return {protoBuffer, ed25519Hex};
-};
-
+/**
+ * Create merge entity object that takes the based db entity and updates it with more recent network retrieved values
+ * @param dbEntity
+ * @param networkEntity
+ * @returns {Object} Updated db entity to insert
+ */
 const getUpdatedEntity = (dbEntity, networkEntity) => {
+  // create duplicate of db entity to update with network values to eventually insert into db
   const updateEntity = {
     ...dbEntity,
   };
 
   let updateNeeded = false;
 
-  // create updated entity to insert to db
   if (dbEntity.auto_renew_period !== networkEntity.autoRenewPeriod.seconds.toString()) {
     updateEntity.auto_renew_period = networkEntity.autoRenewPeriod.seconds.toString();
     updateNeeded = true;
@@ -120,7 +69,7 @@ const getUpdatedEntity = (dbEntity, networkEntity) => {
   }
 
   // mirror t_entities.ed25519_public_key_hex is created based on key so we can use this for both key and hex comparison
-  const {protoBuffer, ed25519Hex} = getBufferAndEd25519HexFromKey(networkEntity.key);
+  const {protoBuffer, ed25519Hex} = utils.getBufferAndEd25519HexFromKey(networkEntity.key);
   if (dbEntity.ed25519_public_key_hex !== ed25519Hex) {
     updateEntity.ed25519_public_key_hex = ed25519Hex;
     updateEntity.key = protoBuffer;
@@ -142,7 +91,6 @@ const getUpdatedEntity = (dbEntity, networkEntity) => {
   }
 
   if (updateNeeded) {
-    // logger.debug(`entity ${dbEntity.id} contained mismatched information`);
     logger.trace(
       `created update entity: ${JSON.stringify(updateEntity)} to replace current db entity ${JSON.stringify(dbEntity)}`
     );
@@ -151,7 +99,12 @@ const getUpdatedEntity = (dbEntity, networkEntity) => {
   return updateNeeded ? updateEntity : null;
 };
 
-const getValidatedEntity = async (csvEntity) => {
+/**
+ * Using csv entity id, compare entity information in mirror db and network state to obtained verified object
+ * @param csvEntity
+ * @returns {Promise<null>}
+ */
+const getVerifiedEntity = async (csvEntity) => {
   let networkEntity;
   try {
     networkEntity = await networkEntityService.getAccountInfo(csvEntity.entity);
@@ -175,6 +128,11 @@ const getValidatedEntity = async (csvEntity) => {
   return getUpdatedEntity(dbEntity, networkEntity);
 };
 
+/**
+ * Retrieve list of objects to update mirror db with. List represent correct information for out of date entities
+ * @param csvEntities
+ * @returns {Promise<unknown[]|[]>}
+ */
 const getUpdateList = async (csvEntities) => {
   logger.info(`Validating entities entities against db entries ...`);
   let updateList = [];
@@ -184,18 +142,23 @@ const getUpdateList = async (csvEntities) => {
 
   const startBalance = await networkEntityService.getAccountBalance();
 
-  updateList = (await Promise.all(csvEntities.map(getValidatedEntity))).filter((x) => !!x);
+  updateList = (await Promise.all(csvEntities.map(getVerifiedEntity))).filter((x) => !!x);
   logger.info(`Update list of length ${updateList.length} retrieved`);
 
   const endBalance = await networkEntityService.getAccountBalance();
-  // logger.info(
-  //   `*** Network accountInfo calls cost ${endBalance.hbars.toTinybars() - startBalance.hbars.toTinybars()} tℏ.
-  //   start: ${startBalance.hbars.toTinybars()} tℏ, end: ${endBalance.hbars.toTinybars()} tℏ`
-  // );
+  logger.info(
+    `*** Network accountInfo calls cost ${endBalance.hbars.toTinybars() - startBalance.hbars.toTinybars()} tℏ.
+    start: ${startBalance.hbars.toTinybars()} tℏ, end: ${endBalance.hbars.toTinybars()} tℏ`
+  );
 
   return updateList;
 };
 
+/**
+ * Update mirror db with list of corrected entities
+ * @param entitiesToUpdate
+ * @returns {Promise<void>}
+ */
 const updateStaleDBEntities = async (entitiesToUpdate) => {
   if (!entitiesToUpdate || entitiesToUpdate.length === 0) {
     logger.info(`No entities to update, skipping update`);
